requirements.txt
celery>=5.5.0
google-genai>=1.15.0
python-dotenv>=1.1.0
redis>=5.2.0
requests>=2.32.0
